Menu Controll
- working on adding elements to menu,
- first level items almost work

+++ app/Http/Controllers/MenuController.php
@@ -21,11 +21,36 @@ class MenuController extends Controller
     $menus=new menu();
 
     foreach($_POST as $id => $oneInput){
-      if($oneInput=='false'){
+      if($id==1){break;}
+
+      elseif($oneInput=='false'){
         $menus->where('id',$id)->delete();
+      }elseif($oneInput=='true'){
+
+      }else{
+        #checking how many elements are in DB and post
+        $sizeDB=count($menus->all());
+        $sizePOST=count($_POST)-1;
+
+        #if element was added then post is bigger than DB
+        if($sizePOST>$sizeDB){
+            for($x=$sizeDB;$x<=$sizePOST;$x++){
+              $menus->insert([
+                'name'=>$_POST[$x][0], 'slug'=>$_POST[$x][1], 'depth'=>'1','parentID'=>'-1'
+              ]);
+              echo '<pre>';
+              var_dump($_POST[$x]);
+              echo '</pre>';
+            }
+        }
+
       }
     }
-    die();
+
+
+    dd($_POST);
+    echo '</br></hr></br>';
+    dd($size);
 
     return redirect($_SERVER['HTTP_REFERER']);
   }
